+++ NEWS.md
@@ -1,6 +1,16 @@
+# RcppThread 0.4.0.9000
+
+# RcppThread 0.5.0.9000
+
+NEW FEATURES
+
+* New functions `parallelFor()` and `ForEach()` allowing parallel `for` loops 
+  with load balancing. Can also be called as method of a `ThreadPool`.
+
 * Options to override `std::thread` and `std::cout` with RcppThread equivalents 
   (preprocessor variables `RCPPTHREAD_OVERRIDE_THREAD` and 
   `RCPPTHREAD_OVERRIDE_COUT`). Both overrides are activated by default.
+
 
 # RcppThread 0.4.0 (June 12, 2018)